filters out unsupported_contract_types_list
